daemon: ceph-disk list replacement tool
ceph-disk list output is really hazardous, so we now have a simpler
scenario to detect journal, block.wal, block.db. It's meant to be used
by ceph-ansible (or other tools) that needs to run the osd activation
stage. For this, it needs to know some variables about the OSD itself.

Signed-off-by: Sébastien Han <seb@redhat.com>